extract classes/interfaces from LoggingController.java
