+++ core/src/main/java/kafka/server/logger/RuntimeLoggerManager.java
@@ -25,7 +25,7 @@ import org.apache.kafka.common.errors.InvalidRequestException;
 import org.apache.kafka.common.message.IncrementalAlterConfigsRequestData.AlterConfigsResource;
 import org.apache.kafka.common.message.IncrementalAlterConfigsRequestData.AlterableConfig;
 import org.apache.kafka.common.protocol.Errors;
-import org.apache.kafka.server.LoggingController;
+import org.apache.kafka.server.logging.LoggingController;
 
 import org.slf4j.Logger;
