+++ core/src/main/scala/kafka/utils/Logging.scala
@@ -18,8 +18,8 @@
 package kafka.utils
 
 import com.typesafe.scalalogging.Logger
+import org.apache.kafka.server.logging.LoggingController
 import org.slf4j.{LoggerFactory, Marker, MarkerFactory}
-import org.apache.kafka.server.LoggingController
 
 object Log4jControllerRegistration {
   private val logger = Logger(this.getClass.getName)
