+++ server/src/main/java/org/apache/kafka/server/logging/NoOpController.java
@@ -1,0 +1,38 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.kafka.server.logging;
+
+import java.util.Collections;
+import java.util.Map;
+
+public class NoOpController extends LoggingControllerDelegate {
+
+    @Override
+    public Map<String, String> loggers() {
+        return Collections.emptyMap();
+    }
+
+    @Override
+    public boolean logLevel(String loggerName, String logLevel) {
+        return false;
+    }
+
+    @Override
+    public boolean unsetLogLevel(String loggerName) {
+        return false;
+    }
+}
